base_script v3.5

+++ install/sh/repo_one_nonreboot.sh
@@ -6,5 +6,13 @@ dnf update -y
 rm /etc/yum.repos.d/*
 cp repo/RedOS-Base.repo /etc/yum.repos.d/ 
 cp repo/RedOS-Updates.repo /etc/yum.repos.d/ 
-reboot
-
+clear
+echo -n "
+Перезагрузить ПК?(да/нет): "
+read askReboot
+if [ $askReboot = "да" ]
+then
+    reboot
+else 
+    ./install.sh
+fi
